requirements.txt
openparse
boto3
flask
python-magic
python-dotenv
langchain
langchain-community
langchain_pinecone
moviepy
openai
pillow
